+++ src/tasks/submit.rs
@@ -1,0 +1,9 @@
+
+pub struct Request {
+    
+}
+impl Request {
+    pub fn new() -> Self {
+        todo!()
+    }
+}